Remove pointless practice pages.
Had some extra pages that I had added for the benefit and joys of learning
But now that must go so I can have a nice website, well nice is subjective.
Maybe simple is a better word at this point..., let's be honest websites are
not my greatest talent...yet lol.